[Feat] OnboardingUseCaseType 구현 #38
